+++ springql-core/src/sql_processor/sql_parser/pest_parser_impl.rs
@@ -1106,6 +1106,7 @@ impl PestParserImpl {
         let s = self_as_str(&mut params);
         match s.to_ascii_uppercase().as_str() {
             "INTEGER" => Ok(SqlType::integer()),
+            "UNSIGNED INTEGER" => Ok(SqlType::unsigned_integer()),
             x => {
                 eprintln!("Unexpected data type parsed: {}", x);
                 unreachable!();
@@ -1216,6 +1217,7 @@ impl PestParserImpl {
         match typ.as_ref() {
             "NET_CLIENT" => Ok(SourceReaderType::NetClient),
             "NET_SERVER" => Ok(SourceReaderType::NetServer),
+            "CAN" => Ok(SourceReaderType::CAN),
             _ => Err(SpringError::Sql(anyhow!(
                 "Invalid source reader name: {}",
                 typ