+++ springql-core/src/stream_engine/autonomous_executor/row.rs
@@ -8,7 +8,7 @@ mod value;
 
 pub use column::StreamColumns;
 pub use column_values::ColumnValues;
-pub use foreign_row::{JsonObject, JsonSourceRow, SourceRow, SourceRowFormat};
+pub use foreign_row::{CANFrameSourceRow, JsonObject, JsonSourceRow, SourceRow, SourceRowFormat};
 pub use rowtime::RowTime;
 pub use value::{NnSqlValue, SpringValue, SqlCompareResult, SqlValue, SqlValueHashKey};